refactor: Remove @flex/params and @flex/middleware packages
