--- libs/middlewares/eslint.config.mjs
@@ -1,3 +1,0 @@
-import { config } from "@flex/config/eslint";
-
-export default config;
