--- libs/middlewares/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "@flex/middlewares",
-  "version": "1.0.0",
-  "type": "module",
-  "private": true,
-  "exports": {
-    ".": "./src/index.ts"
-  },
-  "scripts": {
-    "tsc": "tsc --noEmit",
-    "lint": "eslint --max-warnings=0 .",
-    "test": "vitest --passWithNoTests"
-  },
-  "dependencies": {
-    "@aws-sdk/client-secrets-manager": "3.1032.0",
-    "@flex/utils": "workspace:*",
-    "@middy/core": "7.2.3",
-    "@middy/secrets-manager": "7.2.3"
-  },
-  "devDependencies": {
-    "@flex/config": "workspace:*",
-    "@flex/testing": "workspace:*",
-    "@types/aws-lambda": "8.10.161",
-    "eslint": "10.2.1",
-    "typescript": "6.0.3",
-    "vitest": "4.1.4"
-  }
-}
