--- libs/middlewares/src/index.ts
@@ -1,6 +1,0 @@
-export {
-  type ContextWithUserId,
-  extractUser,
-  type V2Authorizer,
-} from "./extract-user";
-export { createSecretsMiddleware } from "./secrets";
