--- libs/middlewares/vitest.config.ts
@@ -1,3 +1,0 @@
-import { config } from "@flex/config/vitest";
-
-export default config;
