--- libs/params/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "@flex/params",
-  "version": "0.0.0",
-  "private": true,
-  "type": "module",
-  "exports": {
-    ".": "./src/index.ts"
-  },
-  "scripts": {
-    "tsc": "tsc --noEmit",
-    "lint": "eslint --max-warnings=0 .",
-    "test": "vitest --passWithNoTests"
-  },
-  "dependencies": {
-    "@aws-lambda-powertools/parameters": "2.33.0",
-    "@flex/logging": "workspace:*",
-    "zod": "4.3.6"
-  },
-  "devDependencies": {
-    "@flex/config": "workspace:*",
-    "@flex/testing": "workspace:*",
-    "@flex/utils": "workspace:*",
-    "@types/node": "25.6.0",
-    "eslint": "10.2.1",
-    "typescript": "6.0.3",
-    "vitest": "4.1.4"
-  }
-}
